cleaned up ./auth/auth.js, added middleware to allow group-info to provide more details if the caller is logged in with a token and authorized
