+++ src/middleware.js
@@ -353,6 +353,18 @@ const mid = {
 		return next();
 	},
 	// permissions and roles
+	async pubOrContext(req, res,next) {
+		if(!req.user) return next();
+		else {
+			return mid.organizationContext(req, res, next);
+		}
+	},
+	async pubOrPermissions (req, res, next) {
+		if(!req.user) return next();
+		else {
+			return access.permissions(req, res, next);
+		}
+	},
 	permissions: access.permissions,
 	access: access.enforce,
 	setRoleTarget: access.setRoleTarget,
@@ -365,7 +377,8 @@ const mid = {
 	isBasic: authorizer.isBasic,
 	isSimpleIAT: authorizer.isSimpleIAT,
 	isAccessOrSimpleIAT: authorizer.isAccessOrSimpleIAT,
-	isWhitelisted: authorizer.isWhitelisted
+	isWhitelisted: authorizer.isWhitelisted,
+	isPublicOrAuth: authorizer.publicOrAuth
 };
 
 export default mid;
